// src/framework/client/ErrorCode.ts
export enum ErrorCode {
    
    SUCCESS = 0, // success

    /** ts 客户端异常 */
    TS_INTERNAL_ERROR = 99000001, // 内部错误，客户端请求失败
    TS_HTTP_ERROR = 99000002, // HTTP状态码异常
    TS_REQUEST_CANCELLED = 99000003, // 已取消
    TS_UPLOAD_ERR = 99000004, // 上传失败
    TS_PARSER_JSON_ERR = 99000005, // JSON解析错误
    TS_NO_FILE_SEL = 99000006, // 未选择文件
    TS_FILE_PROCESS_ERR = 99000007, // 文件处理异常
    TS_FILE_TOO_BIG = 99000008, // 文件过大

    /** 服务器业务异常 */
    SERVER_COMMON_INTERNAL = 1000, // 未知错误
    SERVER_COMMON_JSON_FORMAT = 1001, // JSON格式错误
    SERVER_COMMON_BUSY = 1002, // 服务器资源不足
    SERVER_COMMON_TASK_NO_EXISTS = 1003, // 任务不存在
    SERVER_COMMON_REQ_TOO_MANY = 1004, // 过多请求
    SERVER_COMMON_TASK_REPEAT = 1005, // 任务重复
    SERVER_COMMON_INSTALL_ID_NULL = 1006, // 安装编号为空
    SERVER_COMMON_USER_NOT_LOGIN = 1008, // 用户未登录
    SERVER_COMMON_LOGIN_INVALID = 1009, // 用户登录已失效
    SERVER_COMMON_TASK_TIMEOUT_RETRY = 1020, // 等待任务结果超时
    SERVER_COMMON_NO_CACHE_HIT = 1021, // 没有命中缓存
    SERVER_COMMON_USAGE_LIMITED = 1024, // API用量达到限制
    SERVER_COMMON_NO_REQUEST_CONTEXT = 1025, // 没有请求上下文
    SERVER_COMMON_INVALID_REQUEST_CONTEXT = 1026, // 非法请求上下文
    SERVER_COMMON_UPGRADE_CLIENT = 1027, // 请升级客户端
    SERVER_COMMON_NEED_VIP = 1028, // 请升级VIP

    /** 用户模块异常 */
    APP_SERVER_LOGIN_SEND_CODE_TOO_MANY = 2006, // 发送验证码过于频繁，请稍后再试
    APP_SERVER_LOGIN_APPID_ERR = 2007, // APPID不存在
    APP_SERVER_LOGIN_AUTH_NET_ERR = 2008, // 登陆验证异常
    APP_SERVER_LOGIN_AUTH_FAIL = 2009, // 登陆验证失败
    APP_SERVER_USER_NAME_TOO_LONG = 2010, // 用户名太长
    APP_SERVER_SEND_EMAIL_VERIFY_CODE_ERR = 2011, // 发送邮件验证码失败
    APP_SERVER_VERIFY_CODE_NON = 2012, // 验证码不存在或已过期
    APP_SERVER_VERIFY_CODE_WRONG = 2013, // 验证码错误
    APP_SERVER_EMAIL_ALREADY_REGISTER = 2014, // 此邮箱已被注册
    APP_SERVER_USERNAME_OR_PASSWORD_WRONG = 2015, // 用户名或密码错误
    APP_SERVER_NEED_OLD_PASSWORD = 2016, // 需要输入旧密码
    APP_SERVER_PASSWORD_INVALID = 2017, // 密码不合法
    
    /** 支付模块异常 */
    APP_SERVER_ALIPAY_APPID_ERR = 3001, // 支付宝APPID错误
    APP_SERVER_ALIPAY_GET_PAYINFO_ERR = 3002, // 获取支付信息失败
    APP_SERVER_SALES_PACKAGE_UNMATCH = 3003, // 套餐不匹配
    APP_SERVER_PAY_METHOD_ERR = 3004, // 支付方式错误
    APP_SERVER_ORDER_ID_NOT_FOUND = 3005, // 找不到订单
    APP_SERVER_ALREADY_VIP_FOREVER = 3006, // 您已经是永久会员，无法继续购买

    NLP_NOT_SUPPORT_TANS_TARGET = 5001, // 不支持的翻译目标语言
    NLP_NO_TEXT_TO_TRANS = 5002, // 没有需要翻译的文本
    NLP_TRANS_FAILED = 5003, // 翻译失败
    NLP_TRANS_AI_REJECT = 5004, // AI拒绝翻译
    NLP_TRANS_TOO_LONG = 5005, // 翻译文本过长
    NLP_NOT_SUPPORT_TANS_SRC = 5006, // 不支持的源语言

}